# docs/swagger.yaml
basePath: /api/v1
consumes:
    - application/json
definitions:
    CreateNoteRequest:
        description: Create a note request
        properties:
            description:
                type: string
                x-go-name: Description
            name:
                type: string
                x-go-name: Name
        required:
            - name
            - description
        type: object
        x-go-package: todo-api-golang/internal/todo/note
    GetNoteResponse:
        properties:
            description:
                type: string
                x-go-name: Description
            id:
                type: string
                x-go-name: ID
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: todo-api-golang/internal/todo/note
    UpdateNoteRequest:
        description: Update a note request
        properties:
            description:
                type: string
                x-go-name: Description
            name:
                type: string
                x-go-name: Name
        required:
            - name
            - description
        type: object
        x-go-package: todo-api-golang/internal/todo/note
host: localhost:8080
info:
    description: '# Documentation for Todo API'
    title: Todo API
    version: 1.0.0
paths:
    /health:
        get:
            description: Check health of the api
            operationId: Health
            responses:
                "200":
                    $ref: '#/responses/helthReponseWrapper'
            summary: Check health of the api
            tags:
                - Health
    /notes:
        get:
            description: Returns a list of notes from the database
            operationId: Notes
            responses:
                "200":
                    $ref: '#/responses/GetNotesResponse'
            summary: Returns a list of notes
            tags:
                - Notes
        post:
            description: Create a new note in a database
            operationId: createNoteRequestWrapper
            parameters:
                - in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/CreateNoteRequest'
            responses:
                "201":
                    $ref: '#/responses/CreateNoteResponse'
                "422":
                    $ref: '#/responses/errorResponseWrapper'
            summary: Creates a new note
            tags:
                - Notes
    /notes/{noteId}:
        get:
            description: Returns a single note from the database
            operationId: noteIdQueryParamWrapper
            parameters:
                - description: The id of the note for which the operation relates
                  in: path
                  name: noteId
                  required: true
                  type: string
                  x-go-name: ID
            responses:
                "200":
                    $ref: '#/responses/GetNoteByIdResponse'
            summary: Returns a single note
            tags:
                - Notes
        patch:
            description: Update a new note in a database
            operationId: updateNoteRequestWrapper
            parameters:
                - description: The id of the note for which the operation relates
                  in: path
                  name: noteId
                  required: true
                  type: string
                  x-go-name: ID
                - in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/UpdateNoteRequest'
            responses:
                "204":
                    $ref: '#/responses/noContentResponseWrapper'
                "422":
                    $ref: '#/responses/errorResponseWrapper'
            summary: Update an existing note
            tags:
                - Notes
produces:
    - application/json
responses:
    CreateNoteResponse:
        description: Create a note response
        headers:
            description:
                type: string
            id:
                type: string
            name:
                type: string
    GetNoteByIdResponse:
        description: A single note returns in the response
        headers:
            description:
                type: string
            id:
                type: string
            name:
                type: string
    GetNotesResponse:
        description: List notes returns
        schema:
            items:
                $ref: '#/definitions/GetNoteResponse'
            type: array
    errorResponseWrapper:
        description: ""
        headers:
            code:
                format: int64
                type: integer
            message:
                type: string
            type:
                type: string
    helthReponseWrapper:
        description: Returns Healthy if the api is working
        schema:
            type: string
    noContentResponseWrapper:
        description: No content is returned by this API endpoint
schemes:
    - http
    - https
swagger: "2.0"